gst: changes the type of gst_aravis->gain from int to double.
Since arv_camera returns double when gain is queried, it makes sense that gain handles this value accordingly.

+++ gst/gstaravis.h
@@ -43,7 +43,7 @@ struct _GstAravis {
 
 	char *camera_name;
 
-	gint64 gain;
+	double gain;
 	gboolean gain_auto;
 	double exposure_time_us;
 	gboolean exposure_auto;
